requirements.txt
hydra-core
pydantic-ai[logfire]
loguru
dotenv